apply @kortschak suggestions
